store fix

+++ store/v2/commitment/store_test_suite.go
@@ -176,7 +176,7 @@ func (s *CommitStoreTestSuite) TestStore_LoadVersion() {
 	for i := uint64(1); i <= latestVersion; i++ {
 		commitInfo, _ := targetStore.GetCommitInfo(i)
 		s.Require().NotNil(commitInfo)
-		s.Require().Equal(i, commitInfo.Version)
+		s.Require().Equal(i, uint64(commitInfo.Version))
 	}
 
 	// rollback to a previous version
